requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.108.0
openai>=1.3.0
chromadb>=0.4.18
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
pydantic>=2.8.0
pydantic-settings>=2.1.0

